Deprecated: Quickly format a parameter table into flextable objects

Use createParameterTable() and printParameterTable() instead.

Usage

formatParameterTable(
  pars,
  include.fix = "ifNotZero",
  include,
  drop,
  drop.symbol = "not used"
)

Arguments

pars

A parameter table as generated by `NMwork::createParameterTable()`.

include.fix

Either a logical or "ifNotZero".

include

See code

drop

See code

drop.symbol

See code

Details

A simple printing function for createParameterTable(). It creates three objects.

partab_full: a df with all columns for selected parameters.

partab_detail: a flextable with selected columns for selected parameters.

partab_present: a flextable with fewer selected columns for selected parameters.

Notes: I suggest dropping the asterisk on parameter.ltx and instead include a nicely formated transformation column in createParameterTable(). It could be included in a parameter table only if non-standard transformations are used.

formatParameterTable is a bad name. This function selects rows and columns in three different levels of detail and returns those three objects. Formatting should happen in `createParameterTable()`.